third_party/blink/web_tests/external/wpt/css/css-fonts/math-script-level-and-math-style/math-script-level-004.tentative-expected.txt

This is a testharness.js-based test.
FAIL scriptPercentScaleDown=80, scriptScriptPercentScaleDown=40 assert_equals: expected "0" but got ""
FAIL scriptPercentScaleDown=0, scriptScriptPercentScaleDown=40 assert_equals: expected "0" but got ""
FAIL scriptPercentScaleDown=80, scriptScriptPercentScaleDown=0 assert_equals: expected "0" but got ""
FAIL No MATH table assert_equals: expected "0" but got ""
Harness: the test ran to completion.
